refactor: split views into separate modules for better organisation
