+++ milk2meat/core/urls.py
@@ -1,25 +1,28 @@
 from django.urls import path
 
-from . import views
+from .views import books as book_views
+from .views import notes as note_views
+from .views import search as search_views
+from .views import tags as tags_views
 
 app_name = "core"
 
 urlpatterns = [
     # Books
-    path("books/", views.BookListView.as_view(), name="book_list"),
-    path("books/<int:pk>/", views.BookDetailView.as_view(), name="book_detail"),
-    path("books/<int:pk>/edit/", views.BookUpdateView.as_view(), name="book_edit"),
+    path("books/", book_views.BookListView.as_view(), name="book_list"),
+    path("books/<int:pk>/", book_views.BookDetailView.as_view(), name="book_detail"),
+    path("books/<int:pk>/edit/", book_views.BookUpdateView.as_view(), name="book_edit"),
     # Notes
-    path("notes/", views.NoteListView.as_view(), name="note_list"),
-    path("notes/create/", views.NoteCreateView.as_view(), name="note_create"),
-    path("notes/<uuid:pk>/", views.NoteDetailView.as_view(), name="note_detail"),
-    path("notes/<uuid:pk>/edit/", views.NoteUpdateView.as_view(), name="note_edit"),
-    path("notes/<uuid:pk>/delete/", views.note_delete_view, name="note_delete"),
+    path("notes/", note_views.NoteListView.as_view(), name="note_list"),
+    path("notes/create/", note_views.NoteCreateView.as_view(), name="note_create"),
+    path("notes/<uuid:pk>/", note_views.NoteDetailView.as_view(), name="note_detail"),
+    path("notes/<uuid:pk>/edit/", note_views.NoteUpdateView.as_view(), name="note_edit"),
+    path("notes/<uuid:pk>/delete/", note_views.note_delete_view, name="note_delete"),
     # Secure file access
-    path("notes/<uuid:note_id>/file/", views.serve_protected_file, name="serve_protected_file"),
+    path("notes/<uuid:note_id>/file/", note_views.serve_protected_file, name="serve_protected_file"),
     # Tags
-    path("tags/", views.TagListView.as_view(), name="tag_list"),
-    path("api/note-types/create/", views.create_note_type_ajax, name="create_note_type_ajax"),
+    path("tags/", tags_views.TagListView.as_view(), name="tag_list"),
+    path("api/note-types/create/", note_views.create_note_type_ajax, name="create_note_type_ajax"),
     # Search
-    path("search/", views.GlobalSearchView.as_view(), name="global_search"),
+    path("search/", search_views.GlobalSearchView.as_view(), name="global_search"),
 ]
