+++ milk2meat/core/views/notes.py
@@ -1,4 +1,3 @@
-import json
 import logging
 
 from django.contrib import messages
@@ -13,84 +12,12 @@ from django.utils.safestring import mark_safe
 from django.views.decorators.http import require_GET, require_POST
 from django.views.generic import CreateView, DetailView, ListView, UpdateView
 from taggit.models import Tag
-from watson import search as watson
 
-from .forms import BookEditForm, NoteForm, NoteTypeForm
-from .models import Book, Note, NoteType, Testament
-from .utils.markdown import parse_markdown
+from ..forms import NoteForm, NoteTypeForm
+from ..models import Book, Note, NoteType
+from ..utils.markdown import parse_markdown
 
 logger = logging.getLogger(__name__)
-
-
-class BookListView(LoginRequiredMixin, ListView):
-    model = Book
-    template_name = "core/book_list.html"
-    context_object_name = "books"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        # Organize books by testament
-        context["old_testament"] = Book.objects.filter(testament=Testament.OT).order_by("number")
-        context["new_testament"] = Book.objects.filter(testament=Testament.NT).order_by("number")
-        return context
-
-
-class BookDetailView(LoginRequiredMixin, DetailView):
-    model = Book
-    template_name = "core/book_detail.html"
-    context_object_name = "book"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        # Parse and render markdown fields
-        book = self.object
-        if book.title_and_author:
-            context["title_and_author_html"] = mark_safe(parse_markdown(book.title_and_author))
-        if book.date_and_occasion:
-            context["date_and_occasion_html"] = mark_safe(parse_markdown(book.date_and_occasion))
-        if book.characteristics_and_themes:
-            context["characteristics_and_themes_html"] = mark_safe(parse_markdown(book.characteristics_and_themes))
-        if book.christ_in_book:
-            context["christ_in_book_html"] = mark_safe(parse_markdown(book.christ_in_book))
-        if book.outline:
-            context["outline_html"] = mark_safe(parse_markdown(book.outline))
-
-        # Add timeline data if it exists
-        if self.object.timeline:
-            context["timeline_data"] = self.object.timeline.get("events", [])
-
-        return context
-
-
-class BookUpdateView(LoginRequiredMixin, UpdateView):
-    model = Book
-    form_class = BookEditForm
-    template_name = "core/book_edit.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        # Add timeline data as JSON string for the form
-        if self.object.timeline:
-            context["timeline_json"] = json.dumps(self.object.timeline)
-        else:
-            context["timeline_json"] = json.dumps({"events": []})
-
-        return context
-
-    def form_valid(self, form):
-        """Process the form data and save"""
-        try:
-            response = super().form_valid(form)
-            messages.success(self.request, f"Successfully updated {self.object.title}.")
-            return response
-        except Exception as e:
-            messages.error(self.request, f"Error saving book: {str(e)}")
-            return super().form_invalid(form)
-
-    def get_success_url(self):
-        return reverse_lazy("core:book_detail", kwargs={"pk": self.object.pk})
 
 
 class NoteListView(LoginRequiredMixin, ListView):
@@ -274,112 +201,6 @@ class NoteUpdateView(LoginRequiredMixin, UpdateView):
 
     def get_success_url(self):
         return reverse_lazy("core:note_detail", kwargs={"pk": self.object.pk})
-
-
-class TagListView(LoginRequiredMixin, ListView):
-    """View for browsing tags"""
-
-    template_name = "core/tag_list.html"
-    context_object_name = "tags"
-
-    def get_queryset(self):
-        """Get all tags used by the current user's notes with counts"""
-        from django.db.models import Count
-        from taggit.models import Tag
-
-        # Get all note IDs for the current user
-        user_note_ids = Note.objects.get_queryset_for_user(self.request.user).values_list("id", flat=True)
-
-        # Get tags with note counts using UUIDTaggedItem
-        return (
-            Tag.objects.filter(
-                core_uuidtaggeditem_items__content_type__app_label="core",
-                core_uuidtaggeditem_items__content_type__model="note",
-                core_uuidtaggeditem_items__object_id__in=user_note_ids,
-            )
-            .annotate(note_count=Count("core_uuidtaggeditem_items"))
-            .order_by("name")
-        )
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        # Organize tags by first letter for the alphabetical section
-        tags_by_letter = {}
-        for tag in context["tags"]:
-            if tag.name:
-                first_letter = tag.name[0].upper()
-                if first_letter not in tags_by_letter:
-                    tags_by_letter[first_letter] = []
-                tags_by_letter[first_letter].append(tag)
-
-        context["tags_by_letter"] = tags_by_letter
-
-        # Get tags with count for the tag cloud (sorted by count)
-        context["tags_for_cloud"] = context["tags"].order_by("-note_count")
-
-        return context
-
-
-class GlobalSearchView(LoginRequiredMixin, ListView):
-    template_name = "core/search_results.html"
-    context_object_name = "search_results"
-    paginate_by = 20
-
-    def get_queryset(self):
-        query = self.request.GET.get("q", "")
-
-        if not query:
-            return []
-
-        # Perform search with watson
-        search_results = watson.search(
-            query,
-            # Limit search to user's own notes
-            models=(
-                Note.objects.get_queryset_for_user(self.request.user),
-                Book.objects.all(),
-            ),
-        )
-
-        return search_results
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        # Add search query to context
-        query = self.request.GET.get("q", "")
-        context["search_query"] = query
-
-        if query:
-            # For accurate result counting, we need to get the total count
-            # from the paginator rather than the page object
-            context["total_count"] = context["paginator"].count
-
-            # Add pagination info
-            page_obj = context["page_obj"]
-            if page_obj.has_next() or page_obj.has_previous():
-                start_index = page_obj.start_index()
-                end_index = page_obj.end_index()
-                context["showing_range"] = {
-                    "start": start_index,
-                    "end": end_index,
-                }
-
-            # Group results by model type (only for current page)
-            if context["search_results"]:
-                from itertools import groupby
-
-                # Group results by model type
-                results_by_type = {}
-                for model_name, group in groupby(
-                    context["search_results"], key=lambda x: x.content_type.model_class().__name__
-                ):
-                    results_by_type[model_name] = list(group)
-
-                context["results_by_type"] = results_by_type
-
-        return context
 
 
 @require_POST
